Task 7: To KV

+++ 0x00-python_variable_annotations/7-to_kv.py
@@ -8,4 +8,4 @@ def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
     """
     Returns a tuple with the string and the square of the number as a float.
     """
-    return (k, float(v * v))
+    return (k, (v * v))
